Stop layout thrashing in drag & drop
When building and updating the animated layer for drag & drop
rearranging, we were thrashing the layout by repeatedly measuring a node
and then moving another node, triggering lots of reflows.  This makes it
very slow when there are a lot of nodes to clone.

Instead, we can gather all the positions first, then set them all.